requirements/local.txt
-r common.txt

# Used for code coverage
coverage==4.5.4

# Used in local testing
WebTest==2.0.21
beautifulsoup4==4.4.1
WebOb==1.6.1

# Used to help with Django related debugging
django-debug-toolbar==2.0
